ms-usuarios/requirements.txt
fastapi[standard]==0.116.1
pydantic==2.11.7
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic-settings==2.10.1
bcrypt==4.3.0
python-dotenv==1.1.1 # Para levantar variables de entorno
hashids==1.3.1  # para hashear

passlib[bcrypt]==1.7.4   #para encriptar contraseñas
# python-jose==3.5.0   #para generar y verificar tokens JWT
python-multipart==0.0.20  #para soportar form data (útil para login)

pyjwt==2.10.1

minio==7.2.16

alembic==1.16.4
pydantic[email]==2.11.7
python-jose[cryptography]==3.5.0
argon2-cffi==25.1.0
